// AngularTest/src/app/Models/hoadon.ts
export interface hoadon {
  maHD: number;
  ngayMua: string;
  tongTien: number;
  tenKhachHang: string;
  code: string;
  thanhPho: string;
  quan_Huyen: string;
  xa_Phuong: string;
  ghi_Chu: string;
  email: string;
  thanh_Toan: number;
  soLuong: number;
  soDienThoai: string;
  status: number;
}